separating the code
