+++ custom_components/schulferien/feiertag_sensor.py
@@ -1,0 +1,95 @@
+"""FeiertagSensor-Modul für die Schulferien-Integration."""
+
+import logging
+from datetime import datetime, timedelta
+import aiohttp
+from homeassistant.helpers.entity import Entity
+from .api_utils import hole_daten, parse_daten
+from .const import API_URL_FEIERTAGE
+
+_LOGGER = logging.getLogger(__name__)
+
+class FeiertagSensor(Entity):
+    """Sensor für Feiertage."""
+
+    def __init__(self, hass, config):
+        self._hass = hass
+        self._name = config["name"]
+        self._land = config["land"]
+        self._region = config["region"]
+        self._last_update_date = None
+        self._heute_feiertag = None
+        self._naechster_feiertag = {"name": None, "datum": None}
+
+    @property
+    def name(self):
+        return self._name
+
+    @property
+    def unique_id(self):
+        return f"sensor.feiertag_{self._land}_{self._region}_{self._name}"
+
+    @property
+    def state(self):
+        return "Feiertag" if self._heute_feiertag else "Kein Feiertag"
+
+    @property
+    def extra_state_attributes(self):
+        return {
+            "Land": self._land,
+            "Region": self._region,
+            "Nächster Feiertag": self._naechster_feiertag["name"],
+            "Datum des nächsten Feiertags": self._naechster_feiertag["datum"],
+        }
+
+    async def async_update(self, session=None):
+        """Aktualisiere den Sensor mit den neuesten Daten von der API."""
+        heute = datetime.now().date()
+        if self._last_update_date == heute:
+            _LOGGER.debug("Die API für Feiertage wurde heute bereits abgefragt.")
+            return
+
+        close_session = False
+        if session is None:
+            session = aiohttp.ClientSession()
+            close_session = True
+
+        try:
+            api_parameter = {
+                "countryIsoCode": self._land,
+                "subdivisionCode": self._region,
+                "validFrom": heute.strftime("%Y-%m-%d"),
+                "validTo": (heute + timedelta(days=365)).strftime("%Y-%m-%d"),
+            }
+
+            feiertage_daten = await hole_daten(API_URL_FEIERTAGE, api_parameter, session)
+            feiertage_liste = parse_daten(feiertage_daten, typ="feiertage")
+
+            self._heute_feiertag = any(
+                feiertag["start_datum"] == heute for feiertag in feiertage_liste
+            )
+
+            zukunft_feiertage = [
+                feiertag for feiertag in feiertage_liste if feiertag["start_datum"] > heute
+            ]
+            if zukunft_feiertage:
+                naechster_feiertag = min(
+                    zukunft_feiertage, key=lambda f: f["start_datum"]
+                )
+                self._naechster_feiertag["name"] = naechster_feiertag["name"]
+                self._naechster_feiertag["datum"] = naechster_feiertag["start_datum"].strftime(
+                    "%d.%m.%Y"
+                )
+            else:
+                self._naechster_feiertag["name"] = None
+                self._naechster_feiertag["datum"] = None
+
+            self._last_update_date = heute
+
+        except RuntimeError as error:
+            _LOGGER.warning("API konnte nicht erreicht werden: %s", error)
+
+        finally:
+            if close_session:
+                await session.close()
+                _LOGGER.debug("Session wurde geschlossen.")
